materialized: modularize http module
It's much easier to keep track of things if handlers are grouped in
thematic submodules. This commit is pure code movement.
