+++ src/materialized/src/http/prof.rs
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-//! Profiling endpoints.
+//! Profiling HTTP endpoints.
 
 use std::io::Read;
 use std::sync::{Arc, Mutex};
